Allow reading completion with ido or similar
This *partially* reverts commit b269f3c, "Simply use completing-read
everywhere", but avoids the original problem that caused that change by not
doing anything when either completing-read-function or completing-read
itself are not default.  Company-quickhelp does the latter. We do this
by manipulating completing-read-function only.

Ido is a good match because the list of guesses is small and the first
is usually right. See also github discussion for commit b269f3c.