Add capability to jump to MRI C method sources
Human-readable explanation of the scope of changes is below:

Overview
--------

This PR adds the ability to (usually!) jump to the C sources of
Ruby methods.

It only works if the user has GNU Global installed and sets a few
Elisp variables to the right values.

If the user doesn't set the variable `robe-use-global-p` to true,
the behavior of Robe stays exactly as it is now -- none of
the "jump to C source" machinery is engaged.

Summary of code changes
-----------------------

- Add functions `robe-list-c-method-files`,
  `robe-determine-c-file-for`, `robe-jump-to-c-source`

- Add variables `robe-use-global-p`,
  `robe-ruby-source-directory`, `robe-global-command`

- Change the interface of `robe-jump-to` to add optional
  arguments that allow it to pass in the value of
  `thing-at-point` for the "jump to C source" machinery.  Because
  the arguments are optional, we don't need to change any of
  `robe-jump-to`'s callers

- Modify `robe-jump-to` to check the value of `robe-use-global-p`
  to decide whether to try to jump to C method sources.  For
  users who haven't set this variable, Robe works exactly as it
  does today.

Remaining bugs/weird UX
-----------------------

- Finding the C source of the method is not 100% accurate - some
  of this is due to GLOBAL, some due to the regex search's
  incomplete coverage

- When jumping back from the C source, you have to hit `M-,`
  twice for some reason (possibly a `ggtags-mode` thing)

- Sometimes you have to enter the module name in the minibuffer
  twice to jump to C-based definitions, I'm not sure why yet